+++ src/main/java/pro/sky/hogwarts/school/repository/StudentRepository.java
@@ -1,13 +1,11 @@
 package pro.sky.hogwarts.school.repository;
 
 
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import pro.sky.hogwarts.school.entity.Student;
 
-import java.awt.print.Pageable;
 import java.util.Collection;
 
 @Repository
@@ -20,7 +18,7 @@ public interface StudentRepository  extends JpaRepository<Student, Long> {
     @Query(value = "SELECT COUNT(s) FROM Student s")
     int getStudentsCount();
     @Query(value = "SELECT AVG(age) FROM Student")
-    int getStudentsAvgAge();
+    double getStudentsAvgAge();
 
     /*@Query (
             value = "SELECT * FROM students ORDER BY id DESC LIMIT :count",
