+++ src/main/java/pro/sky/hogwarts/school/service/StudentAggregationService.java
@@ -7,7 +7,6 @@ import pro.sky.hogwarts.school.entity.Student;
 import pro.sky.hogwarts.school.repository.StudentRepository;
 
 import java.util.Collection;
-import java.util.stream.Collectors;
 
 @Service
 public class StudentAggregationService {
@@ -20,7 +19,7 @@ public class StudentAggregationService {
     public int getCount() {
         return studentRepository.getStudentsCount();
     }
-    public int getAvgAge() {
+    public double getAvgAge() {
         return studentRepository.getStudentsAvgAge();
     }
